Minor fixings

+++ src/app/shared/message/messageCtrl.js
@@ -1,9 +1,13 @@
-angular.module('app.message').controller('MessagesCtrl', ['$rootScope', '$scope', 'messageService', function($rootScope, $scope,messageService) {
+angular.module('app.message').controller('MessagesCtrl', ['$rootScope', '$scope', '$location', 'messageService', function($rootScope, $scope, $location, messageService) {
   $scope.messages = messageService.messages;
   
-  messageService.removeAllMessages();
+  console.log($location.path());
+  if($location.path() !== '/login') messageService.removeAllMessages();
+  console.log(messageService.messages);
+  console.log($rootScope.rootFields.showContent);
 
   $scope.closeMessage = function(index) {
     $scope.messages.splice(index, 1);
   };
+
 }]);
